string/hal_i2s: change type of `i` parameter to unsigned long
JIRA: RTOS-1070

+++ hal/string.h
@@ -46,7 +46,7 @@ extern char *hal_strncpy(char *dest, const char *src, size_t n);
 extern char *hal_strchr(const char *str, int z);
 
 
-extern int hal_i2s(char *prefix, char *s, unsigned int i, unsigned char b, char zero);
+extern int hal_i2s(char *prefix, char *s, unsigned long i, unsigned char b, char zero);
 
 
 #endif
